--- examples/data/australia.csv
@@ -1,205 +1,0 @@
-"","X","time","value"
-"1",1,1991.5,3.526591
-"2",2,1991.58333333333,3.180891
-"3",3,1991.66666666667,3.252221
-"4",4,1991.75,3.611003
-"5",5,1991.83333333333,3.565869
-"6",6,1991.91666666667,4.306371
-"7",7,1992,5.088335
-"8",8,1992.08333333333,2.81452
-"9",9,1992.16666666667,2.985811
-"10",10,1992.25,3.20478
-"11",11,1992.33333333333,3.127578
-"12",12,1992.41666666667,3.270523
-"13",13,1992.5,3.73785082
-"14",14,1992.58333333333,3.55877609
-"15",15,1992.66666666667,3.77720173
-"16",16,1992.75,3.92449042
-"17",17,1992.83333333333,4.38653092
-"18",18,1992.91666666667,5.81054917
-"19",19,1993,6.19206769
-"20",20,1993.08333333333,3.45085699
-"21",21,1993.16666666667,3.77230686
-"22",22,1993.25,3.7343029
-"23",23,1993.33333333333,3.90539892
-"24",24,1993.41666666667,4.04968714
-"25",25,1993.5,4.31556552
-"26",26,1993.58333333333,4.56218455
-"27",27,1993.66666666667,4.60866203
-"28",28,1993.75,4.66785129
-"29",29,1993.83333333333,5.09384145
-"30",30,1993.91666666667,7.1799622
-"31",31,1994,6.73147308
-"32",32,1994.08333333333,3.84127758
-"33",33,1994.16666666667,4.39407557
-"34",34,1994.25,4.07534073
-"35",35,1994.33333333333,4.5406449
-"36",36,1994.41666666667,4.64561508
-"37",37,1994.5,4.75260653
-"38",38,1994.58333333333,5.35060467
-"39",39,1994.66666666667,5.20445484
-"40",40,1994.75,5.3016513
-"41",41,1994.83333333333,5.77374216
-"42",42,1994.91666666667,6.20459348
-"43",43,1995,6.74948382
-"44",44,1995.08333333333,4.21606735
-"45",45,1995.16666666667,4.94934946
-"46",46,1995.25,4.8230449
-"47",47,1995.33333333333,5.19475419
-"48",48,1995.41666666667,5.17078711
-"49",49,1995.5,5.25674157
-"50",50,1995.58333333333,5.85527729
-"51",51,1995.66666666667,5.49072901
-"52",52,1995.75,6.11529323
-"53",53,1995.83333333333,6.08847284
-"54",54,1995.91666666667,7.41659792
-"55",55,1996,8.32945212
-"56",56,1996.08333333333,5.06979585
-"57",57,1996.16666666667,5.26255667
-"58",58,1996.25,5.59712628
-"59",59,1996.33333333333,6.110296
-"60",60,1996.41666666667,5.68916084
-"61",61,1996.5,6.48684933
-"62",62,1996.58333333333,6.30056933
-"63",63,1996.66666666667,6.46747574
-"64",64,1996.75,6.82862939
-"65",65,1996.83333333333,6.64907826
-"66",66,1996.91666666667,8.60693721
-"67",67,1997,8.52447101
-"68",68,1997.08333333333,5.27791837
-"69",69,1997.16666666667,5.71430345
-"70",70,1997.25,6.21452908
-"71",71,1997.33333333333,6.41192919
-"72",72,1997.41666666667,6.66771564
-"73",73,1997.5,7.05083102
-"74",74,1997.58333333333,6.70491861
-"75",75,1997.66666666667,7.25098761
-"76",76,1997.75,7.81973318
-"77",77,1997.83333333333,7.39810106
-"78",78,1997.91666666667,10.09623339
-"79",79,1998,8.79851303
-"80",80,1998.08333333333,5.91826076
-"81",81,1998.16666666667,6.53449298
-"82",82,1998.25,6.67573561
-"83",83,1998.33333333333,7.06420058
-"84",84,1998.41666666667,7.38338118
-"85",85,1998.5,7.81349587
-"86",86,1998.58333333333,7.43189221
-"87",87,1998.66666666667,8.27511722
-"88",88,1998.75,8.26044138
-"89",89,1998.83333333333,8.59615575
-"90",90,1998.91666666667,10.55893902
-"91",91,1999,10.3914156
-"92",92,1999.08333333333,6.42153456
-"93",93,1999.16666666667,8.06261937
-"94",94,1999.25,7.2977391
-"95",95,1999.33333333333,7.93691594
-"96",96,1999.41666666667,8.16532298
-"97",97,1999.5,8.71742046
-"98",98,1999.58333333333,9.07096378
-"99",99,1999.66666666667,9.17711337
-"100",100,1999.75,9.25188674
-"101",101,1999.83333333333,9.93313643
-"102",102,1999.91666666667,11.53297428
-"103",103,2000,12.51146235
-"104",104,2000.08333333333,7.45719853
-"105",105,2000.16666666667,8.59119084
-"106",106,2000.25,8.47400037
-"107",107,2000.33333333333,9.3868026
-"108",108,2000.41666666667,9.56039945
-"109",109,2000.5,10.8342948
-"110",110,2000.58333333333,10.64375083
-"111",111,2000.66666666667,9.90816186
-"112",112,2000.75,11.7100413
-"113",113,2000.83333333333,11.34015074
-"114",114,2000.91666666667,12.07913184
-"115",115,2001,14.49758109
-"116",116,2001.08333333333,8.04927477
-"117",117,2001.16666666667,10.31289116
-"118",118,2001.25,9.75335821
-"119",119,2001.33333333333,10.85038183
-"120",120,2001.41666666667,9.96171916
-"121",121,2001.5,11.44360144
-"122",122,2001.58333333333,11.65923889
-"123",123,2001.66666666667,10.64705979
-"124",124,2001.75,12.65213444
-"125",125,2001.83333333333,13.67446631
-"126",126,2001.91666666667,12.96573509
-"127",127,2002,16.30026927
-"128",128,2002.08333333333,9.05348536
-"129",129,2002.16666666667,10.00244879
-"130",130,2002.25,10.78875007
-"131",131,2002.33333333333,12.10670515
-"132",132,2002.41666666667,10.95410107
-"133",133,2002.5,12.84456587
-"134",134,2002.58333333333,12.19649985
-"135",135,2002.66666666667,12.85474787
-"136",136,2002.75,13.54200436
-"137",137,2002.83333333333,13.28764
-"138",138,2002.91666666667,15.13491784
-"139",139,2003,16.82834968
-"140",140,2003.08333333333,9.80021461
-"141",141,2003.16666666667,10.81699371
-"142",142,2003.25,10.65422256
-"143",143,2003.33333333333,12.51232269
-"144",144,2003.41666666667,12.16120969
-"145",145,2003.5,12.9980462
-"146",146,2003.58333333333,12.51727568
-"147",147,2003.66666666667,13.26865761
-"148",148,2003.75,14.73362169
-"149",149,2003.83333333333,13.66938174
-"150",150,2003.91666666667,16.50396561
-"151",151,2004,18.00376795
-"152",152,2004.08333333333,11.93802987
-"153",153,2004.16666666667,12.9979001
-"154",154,2004.25,12.88264507
-"155",155,2004.33333333333,13.94344681
-"156",156,2004.41666666667,13.9894722
-"157",157,2004.5,15.33909742
-"158",158,2004.58333333333,15.37076394
-"159",159,2004.66666666667,16.142005
-"160",160,2004.75,16.685754
-"161",161,2004.83333333333,17.636728
-"162",162,2004.91666666667,18.869325
-"163",163,2005,20.778723
-"164",164,2005.08333333333,12.154552
-"165",165,2005.16666666667,13.402392
-"166",166,2005.25,14.459239
-"167",167,2005.33333333333,14.795102
-"168",168,2005.41666666667,15.705248
-"169",169,2005.5,15.82955
-"170",170,2005.58333333333,17.554701
-"171",171,2005.66666666667,18.100864
-"172",172,2005.75,17.496668
-"173",173,2005.83333333333,19.347265
-"174",174,2005.91666666667,20.031291
-"175",175,2006,23.486694
-"176",176,2006.08333333333,12.536987
-"177",177,2006.16666666667,15.467018
-"178",178,2006.25,14.233539
-"179",179,2006.33333333333,17.783058
-"180",180,2006.41666666667,16.291602
-"181",181,2006.5,16.980282
-"182",182,2006.58333333333,18.612189
-"183",183,2006.66666666667,16.623343
-"184",184,2006.75,21.430241
-"185",185,2006.83333333333,23.575517
-"186",186,2006.91666666667,23.334206
-"187",187,2007,28.038383
-"188",188,2007.08333333333,16.763869
-"189",189,2007.16666666667,19.792754
-"190",190,2007.25,16.427305
-"191",191,2007.33333333333,21.000742
-"192",192,2007.41666666667,20.681002
-"193",193,2007.5,21.83489
-"194",194,2007.58333333333,23.93020353
-"195",195,2007.66666666667,22.93035694
-"196",196,2007.75,23.26333992
-"197",197,2007.83333333333,25.25003022
-"198",198,2007.91666666667,25.80609
-"199",199,2008,29.665356
-"200",200,2008.08333333333,21.654285
-"201",201,2008.16666666667,18.264945
-"202",202,2008.25,23.107677
-"203",203,2008.33333333333,22.91251
-"204",204,2008.41666666667,19.43174
