+++ examples/data/ts/australia.csv
@@ -1,0 +1,205 @@
+idx,value
+0,3.526591
+1,3.180891
+2,3.252221
+3,3.611003
+4,3.565869
+5,4.306371
+6,5.088335
+7,2.81452
+8,2.985811
+9,3.20478
+10,3.127578
+11,3.270523
+12,3.73785082
+13,3.55877609
+14,3.77720173
+15,3.92449042
+16,4.38653092
+17,5.81054917
+18,6.19206769
+19,3.45085699
+20,3.77230686
+21,3.7343029
+22,3.90539892
+23,4.04968714
+24,4.31556552
+25,4.56218455
+26,4.60866203
+27,4.66785129
+28,5.09384145
+29,7.1799622
+30,6.73147308
+31,3.84127758
+32,4.39407557
+33,4.07534073
+34,4.5406449
+35,4.64561508
+36,4.75260653
+37,5.35060467
+38,5.20445484
+39,5.3016513
+40,5.77374216
+41,6.20459348
+42,6.74948382
+43,4.21606735
+44,4.94934946
+45,4.8230449
+46,5.19475419
+47,5.17078711
+48,5.25674157
+49,5.85527729
+50,5.49072901
+51,6.11529323
+52,6.08847284
+53,7.41659792
+54,8.32945212
+55,5.06979585
+56,5.26255667
+57,5.59712628
+58,6.110296
+59,5.68916084
+60,6.48684933
+61,6.30056933
+62,6.46747574
+63,6.82862939
+64,6.64907826
+65,8.60693721
+66,8.52447101
+67,5.27791837
+68,5.71430345
+69,6.21452908
+70,6.41192919
+71,6.66771564
+72,7.05083102
+73,6.70491861
+74,7.25098761
+75,7.81973318
+76,7.39810106
+77,10.09623339
+78,8.79851303
+79,5.91826076
+80,6.53449298
+81,6.67573561
+82,7.06420058
+83,7.38338118
+84,7.81349587
+85,7.43189221
+86,8.27511722
+87,8.26044138
+88,8.59615575
+89,10.55893902
+90,10.3914156
+91,6.42153456
+92,8.06261937
+93,7.2977391
+94,7.93691594
+95,8.16532298
+96,8.71742046
+97,9.07096378
+98,9.17711337
+99,9.25188674
+100,9.93313643
+101,11.53297428
+102,12.51146235
+103,7.45719853
+104,8.59119084
+105,8.47400037
+106,9.3868026
+107,9.56039945
+108,10.8342948
+109,10.64375083
+110,9.90816186
+111,11.7100413
+112,11.34015074
+113,12.07913184
+114,14.49758109
+115,8.04927477
+116,10.31289116
+117,9.75335821
+118,10.85038183
+119,9.96171916
+120,11.44360144
+121,11.65923889
+122,10.64705979
+123,12.65213444
+124,13.67446631
+125,12.96573509
+126,16.30026927
+127,9.05348536
+128,10.00244879
+129,10.78875007
+130,12.10670515
+131,10.95410107
+132,12.84456587
+133,12.19649985
+134,12.85474787
+135,13.54200436
+136,13.28764
+137,15.13491784
+138,16.82834968
+139,9.80021461
+140,10.81699371
+141,10.65422256
+142,12.51232269
+143,12.16120969
+144,12.9980462
+145,12.51727568
+146,13.26865761
+147,14.73362169
+148,13.66938174
+149,16.50396561
+150,18.00376795
+151,11.93802987
+152,12.9979001
+153,12.88264507
+154,13.94344681
+155,13.9894722
+156,15.33909742
+157,15.37076394
+158,16.142005
+159,16.685754
+160,17.636728
+161,18.869325
+162,20.778723
+163,12.154552
+164,13.402392
+165,14.459239
+166,14.795102
+167,15.705248
+168,15.82955
+169,17.554701
+170,18.100864
+171,17.496668
+172,19.347265
+173,20.031291
+174,23.486694
+175,12.536987
+176,15.467018
+177,14.233539
+178,17.783058
+179,16.291602
+180,16.980282
+181,18.612189
+182,16.623343
+183,21.430241
+184,23.575517
+185,23.334206
+186,28.038383
+187,16.763869
+188,19.792754
+189,16.427305
+190,21.000742
+191,20.681002
+192,21.83489
+193,23.93020353
+194,22.93035694
+195,23.26333992
+196,25.25003022
+197,25.80609
+198,29.665356
+199,21.654285
+200,18.264945
+201,23.107677
+202,22.91251
+203,19.43174
