+++ examples/data/ts/beer.csv
@@ -1,4 +1,4 @@
-datetime,value
+idx,value
 1956-01,93.2
 1956-02,96
 1956-03,95.2
