+++ examples/data/ts/salaries.csv
@@ -1,4 +1,4 @@
-date,target
+idx,value
 2016-10-07,6307.0
 2016-10-08,91026.59000000001
 2016-10-09,2446.15
