+++ examples/data/ts/stackoverflow.csv
@@ -1,4 +1,4 @@
-datetime,value
+idx,value
 2009-01-01,832.0
 2009-02-01,893.0
 2009-03-01,1001.0
