+++ examples/time_series/gapfilling.py
