+++ examples/time_series/ts_clstm.py
@@ -3,17 +3,18 @@ import os
 
 import numpy as np
 import pandas as pd
+from sklearn.metrics import mean_squared_error, mean_absolute_error
 
 from examples.pipeline_import_export import create_correct_path
-from examples.time_series.ts_forecasting_composing import prepare_train_test_input, fit_predict_for_pipeline, \
-    display_validation_metric
+from examples.time_series.composing_ts_pipelines import visualise
 from fedot.core.data.data import InputData
 from fedot.core.data.data_split import train_test_data_setup
-from fedot.core.pipelines.node import PrimaryNode, SecondaryNode
+from fedot.core.pipelines.node import PrimaryNode
 from fedot.core.pipelines.pipeline import Pipeline
 from fedot.core.repository.dataset_types import DataTypesEnum
 from fedot.core.repository.tasks import Task, TaskTypesEnum, TsForecastingParams
 from fedot.core.utils import fedot_project_root
+
 
 
 def get_ts_data_long(n_steps=80, forecast_length=5):
@@ -24,7 +25,7 @@ def get_ts_data_long(n_steps=80, forecast_length=5):
     :param forecast_length: the length of forecast
     """
     project_root_path = str(fedot_project_root())
-    file_path = os.path.join(project_root_path, 'examples/data/ts_long.csv')
+    file_path = os.path.join(project_root_path, 'examples/data/ts/ts_long.csv')
     df = pd.read_csv(file_path)
     df = df[df["series_id"] == "traffic_volume"]
     time_series = np.array(df['value'])[:n_steps]
@@ -87,78 +88,34 @@ def clstm_forecasting():
     prediction = np.array(predicted_output.predict[0])
     print(f'Prediction from pipeline loaded from dict {prediction[:4]}')
 
-    display_validation_metric(
-        np.ravel(prediction_before_export),
-        test_data.target, np.concatenate([test_data.features[-window_size:], test_data.target]),
-        True)
+    plot_info = [
+        {'idx': np.arange(np.concatenate([test_data.features, test_data.target]).shape[0]),
+         'series': np.concatenate([test_data.features, test_data.target]),
+         'label': 'Actual time series'
+         },
+        {'idx': np.arange(test_data.idx[0], test_data.idx[0]+horizon),
+         'series': np.ravel(prediction_before_export),
+         'label': 'prediction'
+         },
+        {'idx': [np.arange(test_data.idx[0] + 1)[-1], np.arange(test_data.idx[0] + 1)[-1]],
+         'series': [
+             min(np.concatenate([np.ravel(np.concatenate([test_data.features, test_data.target])),
+                                 np.ravel(prediction_before_export)])),
+             max(np.concatenate([np.ravel(np.concatenate([test_data.features, test_data.target])),
+                                 np.ravel(prediction_before_export)]))
+         ],
+         'label': 'End of Test',
+         'color': 'black'
+         }
+    ]
 
+    rmse = mean_squared_error(test_data.target, prediction_before_export, squared=False)
+    mae = mean_absolute_error(test_data.target, prediction_before_export)
+    print(f'RMSE - {rmse:.4f}')
+    print(f'MAE - {mae:.4f}')
 
-def get_source_pipeline_clstm():
-    """
-    Return pipeline with the following structure:
-    lagged - ridge \
-                    -> ridge
-    clstm - - - - /
-    """
-
-    # First level
-    node_lagged_1 = PrimaryNode('lagged')
-
-    # Second level
-    node_ridge_1 = SecondaryNode('ridge', nodes_from=[node_lagged_1])
-    node_clstm = PrimaryNode('clstm')
-    node_clstm.custom_params = {
-        'window_size': 29,
-        'hidden_size': 50,
-        'learning_rate': 0.004,
-        'cnn1_kernel_size': 5,
-        'cnn1_output_size': 32,
-        'cnn2_kernel_size': 4,
-        'cnn2_output_size': 32,
-        'batch_size': 64,
-        'num_epochs': 3
-    }
-    # Third level - root node
-    node_final = SecondaryNode('ridge', nodes_from=[node_ridge_1, node_clstm])
-    pipeline = Pipeline(node_final)
-
-    return pipeline
-
-
-def run_ts_forecasting_problem(forecast_length=50,
-                               with_visualisation=True) -> None:
-    """ Function launch time series task with composing
-
-    :param forecast_length: length of the forecast
-    :param with_visualisation: is it needed to show the plots
-    """
-    file_path = '../../cases/data/metocean/metocean_data_test.csv'
-
-    df = pd.read_csv(file_path)
-    time_series = np.array(df['sea_height'])
-
-    # Train/test split
-    train_part = time_series[len(time_series)-200:-forecast_length]
-    test_part = time_series[-forecast_length:]
-
-    # Prepare data for train and test
-    train_input, predict_input, task = prepare_train_test_input(train_part,
-                                                                forecast_length)
-
-    # Get pipeline with pre-defined structure
-    init_pipeline = get_source_pipeline_clstm()
-
-    # Init check
-    preds = fit_predict_for_pipeline(pipeline=init_pipeline,
-                                     train_input=train_input,
-                                     predict_input=predict_input)
-
-    display_validation_metric(predicted=preds,
-                              real=test_part,
-                              actual_values=time_series[-100:],
-                              is_visualise=with_visualisation)
+    visualise(plot_info)
 
 
 if __name__ == '__main__':
-    run_ts_forecasting_problem(forecast_length=50,
-                               with_visualisation=True)
+    clstm_forecasting()
