+++ test/unit/data_operations/test_data_operations_implementations.py
@@ -3,7 +3,7 @@ from itertools import product
 import numpy as np
 from examples.classification_with_tuning_example import get_classification_dataset
 from examples.regression_with_tuning_example import get_regression_dataset
-from examples.time_series.ts_gapfilling_example import generate_synthetic_data
+from examples.time_series.gapfilling import generate_synthetic_data
 from fedot.core.data.data import InputData
 from fedot.core.data.data_split import train_test_data_setup
 from fedot.core.data.supplementary_data import SupplementaryData
